Handling Numebrs
Javanese numbers are using symbol that similar with one of javanese letters.
unfortunately, the unicode block for javanese is not including range of numbers,
hence the program will not work as expected when using numbers.

To overcome this, I'm adopting the writing rules for javanese texts.
according to Kongres Aksara Jawa I, one suppose to put numbers between symbol "꧇"

so for example, if you're using this programming language and want to write
1 + 1
you should write it as follow:

꧇꧑꧇ + ꧇꧑꧇

I haven't tested thoroughly, but I hope this help fix some issues.
In the future, I really hope that javanese numbers got their own unicode block.